test_requirements.txt
pytest>=6
pytest-cov
pytest-asyncio>0.11
nest-asyncio
pytest-xdist
pytest-timeout
aiohttp
websockets
pytest-benchmark
ipykernel
pymemfd; platform_system=="Linux"
python-prctl; platform_system=="Linux"
libertem[bqplot]
flask
